Refaktoroitu koodia parempaan järjestykseen
