+++ src/app/lib/constants.ts
@@ -1,0 +1,5 @@
+export const PAGE_SIZES = [10, 20, 30, 50, 100];
+
+export const DEFAULT_PAGE_SIZE = 30;
+
+export const HAKU_SEARCH_PHRASE_DEBOUNCE_DELAY = 500;
